fix: iam config, change variabel name on main, add validation on delete failed job

+++ lambda/dataPortal/admin_dportal_functions.py
@@ -20,7 +20,7 @@ DPORTAL_BUCKET = os.environ.get("DPORTAL_BUCKET")
 ATHENA_METADATA_BUCKET = os.environ.get("ATHENA_METADATA_BUCKET")
 SUBMIT_LAMBDA = os.environ.get("SUBMIT_LAMBDA")
 INDEXER_LAMBDA = os.environ.get("INDEXER_LAMBDA")
-TEMP_BUCKET = os.environ.get("SVEP_TEMP_ARN") 
+TEMP_BUCKET = os.environ.get("SVEP_TEMP_NAME") 
 
 #
 # Files' Admin Functions
@@ -398,21 +398,37 @@ def index_sbeacon(event, context):
 
 
 @router.attach(
-    "/dportal/projects/{project}/clinical-workflows/delete-job/{job_id}",
+    "/dportal/projects/{project}/clinical-workflows/{job_id}",
     "delete",
 )
 def delete_jobid(event, context):
-    DPORTAL_BUCKET_TEMP = match = re.search(r"(svep-[\w\-]+)", TEMP_BUCKET).group(1)
     selectedJOB = event["pathParameters"]["job_id"]
     project_name = event["pathParameters"]["project"]
+    sub = event["requestContext"]["authorizer"]["claims"]["sub"]
     
     try:
-        job = ClinicJobs.get(selectedJOB)
+        #check is user registered in project
+        ProjectUsers.get(project_name, sub) 
+
+        job = ClinicJobs.get(selectedJOB) 
+        if job.job_status.lower() != "failed":
+            return {
+                "success": False,
+                "message": f"Job {selectedJOB} is not in a failed status.",
+            }
         job.delete()
         # delete file from temp data 
-        keys = list_s3_prefix(DPORTAL_BUCKET_TEMP, selectedJOB)
-        deleted = delete_s3_objects(DPORTAL_BUCKET_TEMP, keys)
+        keys = list_s3_prefix(TEMP_BUCKET, selectedJOB)
+        delete_s3_objects(TEMP_BUCKET, keys)
     except ClinicJobs.DoesNotExist:
-        raise PortalError(404, f"Job with ID {selectedJOB} not found")
+        return {
+            "success": False,
+            "message": f"Job with ID {selectedJOB} not found",
+        }
+    except ProjectUsers.DoesNotExist:
+        return {
+            "success": False,
+            "message": "User not registered in project.",
+        }
 
     return {"success": True, "message": f"Deleted job {selectedJOB} from project {project_name}"}
